pyproject.toml
[project]
name = "university"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "drf-yasg>=1.21.8",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
